requirements.txt
flask
opencv-python
picamera
Pillow
numpy